requirements.txt
aiogram
sqlalchemy
asyncpg
python-dotenv
sentence-transformers
qdrant-client
pdfplumber
python-docx
openai
langchain-text-splitters